requirements.txt
python-dotenv
pandas
kafka-python==2.0.2
confluent-kafka==2.3.0
redis==5.0.1
flask==3.0.0
flask-socketio==5.3.6
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
tldextract==5.1.1
python-dotenv==1.0.0


